gui: move config loading into its own class
